# server.js
const express = require("express");
const swaggerUi = require("swagger-ui-express");
const openapiSpecification = require("./openapi.json");
require("dotenv").config();
const { initializeDatabase, pool } = require("./taskRepository");
const supabase = require("./supabaseClient");
const { TriageInputSchema } = require("./src/llm/schema");
const { LlmGatewayError, triageText } = require("./src/llm/triage");

const app = express();
const PORT = 3000;

app.use(express.json());
app.use("/docs", swaggerUi.serve, swaggerUi.setup(openapiSpecification));

app.get("/", (req, res) => {
  res.json({
    name: "Task API",
    version: "1.0",
    endpoints: ["/tasks"]
  });
});

app.get("/health", (req, res) => {
  res.json({ status: "ok" });
});

app.post("/triage", async (req, res) => {
  const input = TriageInputSchema.safeParse(req.body);

  if (!input.success) {
    const message = input.error.issues
      .map((issue) => `${issue.path.join(".") || "request"}: ${issue.message}`)
      .join("; ");
    return res.status(400).json({ error: message });
  }

  try {
    const result = await triageText(input.data.text);
    if (result.kind === "unprocessable") return res.status(422).json({ error: result.error });
    return res.json(result.output);
  } catch (error) {
    if (error instanceof LlmGatewayError && error.timeout) {
      return res.status(504).json({ error: "The LLM provider timed out after 30 seconds." });
    }
    if (error instanceof LlmGatewayError) {
      return res.status(502).json({ error: "The LLM provider could not process this request." });
    }
    console.error("Unexpected triage error:", error.message);
    return res.status(500).json({ error: "Could not triage this message." });
  }
});

app.post("/auth/signup", async (req, res) => {
  const { email, password } = req.body;

  if (!email || !password) {
    return res.status(400).json({ error: "Email and password are required" });
  }

  const { data, error } = await supabase.auth.signUp({ email, password });

  if (error) {
    return res.status(400).json({ error: error.message });
  }

  res.status(201).json({ user: data.user });
});

app.post("/auth/login", async (req, res) => {
  const { email, password } = req.body;

  if (!email || !password) {
    return res.status(400).json({ error: "Email and password are required" });
  }

  const { data, error } = await supabase.auth.signInWithPassword({ email, password });

  if (error) {
    return res.status(401).json({ error: "Invalid login credentials" });
  }

  res.json({
    access_token: data.session.access_token,
    refresh_token: data.session.refresh_token
  });
});

app.get("/public/info", (req, res) => {
  res.json({ message: "Welcome stranger! This info is public." });
});

async function requirePresentedToken(req, res, next) {
  const authorization = req.headers.authorization;

  if (!authorization || !authorization.startsWith("Bearer ")) {
    return res.status(401).json({ error: "Access token required" });
  }

  const token = authorization.slice("Bearer ".length).trim();

  if (!token) {
    return res.status(401).json({ error: "Access token required" });
  }

  const { data, error } = await supabase.auth.getUser(token);

  if (error || !data.user) {
    return res.status(401).json({ error: "Invalid or expired token" });
  }

  req.user = data.user;
  next();
}

app.get("/protected/profile", requirePresentedToken, (req, res) => {
  res.json({
    id: req.user.id,
    email: req.user.email,
    created_at: req.user.created_at
  });
});

app.get("/protected/dashboard", requirePresentedToken, (req, res) => {
  res.json({ message: `Welcome to your dashboard, ${req.user.email}.` });
});

app.post("/auth/logout", requirePresentedToken, async (req, res) => {
  const { error } = await supabase.auth.signOut({ scope: "local" });

  if (error) {
    return res.status(401).json({ error: "Could not log out" });
  }

  res.status(204).send();
});

app.get("/tasks", async (req, res) => {
  const result = await pool.query("SELECT * FROM tasks ORDER BY id");
  res.json(result.rows);
});

app.get("/tasks/:id", async (req, res) => {
  const id = Number(req.params.id);
  const result = await pool.query("SELECT * FROM tasks WHERE id = $1", [id]);
  const task = result.rows[0];

  if (!task) {
    return res.status(404).json({ error: `Task ${req.params.id} not found` });
  }

  res.json(task);
});

app.post("/tasks", async (req, res) => {
  const { title } = req.body;

  if (typeof title !== "string" || title.trim() === "") {
    return res.status(400).json({ error: "Title is required and must be a non-empty string" });
  }

  const result = await pool.query(
    "INSERT INTO tasks (title, done) VALUES ($1, $2) RETURNING *",
    [title.trim(), false]
  );
  const newTask = result.rows[0];

  res.status(201).json(newTask);
});

app.put("/tasks/:id", async (req, res) => {
  const id = Number(req.params.id);
  const taskResult = await pool.query("SELECT * FROM tasks WHERE id = $1", [id]);
  const task = taskResult.rows[0];

  if (!task) {
    return res.status(404).json({ error: `Task ${req.params.id} not found` });
  }

  const { title, done } = req.body;
  const hasTitle = Object.prototype.hasOwnProperty.call(req.body, "title");
  const hasDone = Object.prototype.hasOwnProperty.call(req.body, "done");

  if (!hasTitle && !hasDone) {
    return res.status(400).json({ error: "Provide a title or done value to update" });
  }

  if (hasTitle && (typeof title !== "string" || title.trim() === "")) {
    return res.status(400).json({ error: "Title must be a non-empty string" });
  }

  if (hasDone && typeof done !== "boolean") {
    return res.status(400).json({ error: "Done must be a boolean" });
  }

  const updatedTitle = hasTitle ? title.trim() : task.title;
  const updatedDone = hasDone ? done : task.done;
  const updatedResult = await pool.query(
    "UPDATE tasks SET title = $1, done = $2 WHERE id = $3 RETURNING *",
    [updatedTitle, updatedDone, id]
  );
  const updatedTask = updatedResult.rows[0];

  res.json(updatedTask);
});

app.delete("/tasks/:id", async (req, res) => {
  const id = Number(req.params.id);
  const result = await pool.query("DELETE FROM tasks WHERE id = $1", [id]);

  if (result.rowCount === 0) {
    return res.status(404).json({ error: `Task ${req.params.id} not found` });
  }

  res.status(204).send();
});

async function startServer() {
  await initializeDatabase();
  app.listen(PORT, () => {
    console.log(`Server is running on http://localhost:${PORT}`);
  });
}

if (require.main === module) {
  startServer().catch((error) => {
    console.error("Could not connect to PostgreSQL:", error.message);
    process.exit(1);
  });
}

module.exports = app;